+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 
 ## Fixed
+- Fixed broken method_showcase_for (frozen string error).
 
 ## [2.0.2] - 2020-01-14